refactor(dashboard): extract keeper detail shell
